requirements.txt
torch
transformers
datasets
peft
bitsandbytes
accelerate
sentencepiece
Pillow